# Store already known not parenthesized arrow expression positions for faster exit in case of deep parsing