… tests/baselines/reference/parsingDeepParenthensizedExpression.symbols
… tests/baselines/reference/parsingDeepParenthensizedExpression.types
… tests/cases/compiler/parsingDeepParenthensizedExpression.ts