Fixes #31987

With this change the `polyv.js` attached in the issue, parses in 0.87s while without this change its still not completed after 10min.

```
c:\temp\p>node c:\TypeScript\built\local\tsc.js -allowjs --noEmit polyv.js --extendedDiagnostics
Files:                          6
Lines:                      24830
Nodes:                     366094
Identifiers:                83489
Symbols:                    32095
Types:                       8266
Memory used:              107231K
Assignability cache size:   33122
Identity cache size:            2
Subtype cache size:             0
I/O Read time:              0.01s
Parse time:                 0.87s
Program time:               0.90s
Bind time:                  0.51s
Check time:                 1.34s
Total time:                 2.75s
```